backend/frontend/frontend/src/telegram_bot/requirements.txt
python-telegram-bot==20.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.25.1